--- src/main/java/AddTask.java
@@ -1,6 +1,0 @@
-public class AddTask implements Command {
-    public void execute(ChatSession session, String commandArgs) {
-        session.taskList.add(new Task(commandArgs));
-        session.printMessage(String.format("added: %s", commandArgs));
-    }
-}
