+++ src/main/java/duke/TaskList.java
@@ -1,8 +1,11 @@
+package duke;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+
+import duke.tasks.Task;
 
 public class TaskList {
     public List<Task> internalList;
