+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 public class Ui {
     private ChatSession session;
