+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,7 @@
+package duke.commands;
+import duke.ChatSession;
+import duke.exceptions.InvalidParametersException;
+
 public interface Command {
     public void execute(ChatSession chatSession, String commandArgs) throws InvalidParametersException;
 }
