+++ src/main/java/duke/commands/Mark.java
@@ -1,3 +1,7 @@
+package duke.commands;
+import duke.ChatSession;
+import duke.tasks.Task;
+
 public class Mark implements NamedCommand {
     public String getName() { return "mark"; }
     public void execute(ChatSession session, String commandArgs) {
