+++ src/main/java/duke/commands/NamedCommand.java
@@ -1,3 +1,5 @@
+package duke.commands;
+
 public interface NamedCommand extends Command {
     public String getName();
 }
