+++ src/main/java/duke/commands/Unmark.java
@@ -1,3 +1,7 @@
+package duke.commands;
+import duke.ChatSession;
+import duke.tasks.Task;
+
 public class Unmark implements NamedCommand {
     public String getName() { return "unmark"; }
     public void execute(ChatSession session, String commandArgs) {
