+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,3 +1,4 @@
+package duke.exceptions;
 public class DukeException extends Exception {
     DukeException(String errorMessage) {
         super(errorMessage);
