+++ src/main/java/duke/exceptions/UnrecognizedCommandException.java
@@ -1,3 +1,4 @@
+package duke.exceptions;
 public class UnrecognizedCommandException extends Exception {
     UnrecognizedCommandException(String commandName) {
         super(commandName);
