+++ src/main/java/duke/tasks/ToDo.java
@@ -1,9 +1,10 @@
+package duke.tasks;
 public class ToDo extends Task {
-    ToDo(String name, boolean done) {
+    public ToDo(String name, boolean done) {
         super(name, done);
     }
 
-    ToDo(String name) {
+    public ToDo(String name) {
         this(name, false);
     }
